Add files via upload
Introduction
The purpose of this project is to understand what makes a user default a loan from LendingClub. We will layout the most influencing parameters by considering both the statistical and lexicon based NLP modeling for the prediction.

'Part I' deals with the EDA and traditional ML modeling with ANN. Following are the Algorithms that we've considered:

1)Logistics Regression
2)Decision Tree
3)Random Forest
4)Support Vector Machines
5)Single Layered NN
6)Multi-Layered NN
7)Gradient Boosting
8)Bagging Classifier
9)Bagging Classiefier with Decision Tree Classifier
10)Voting Classifier
11)AdaBoost Classifier
12)Extra Tree Classifier
'Part II' is about the NLP approach to find Default predictions by using text data from the Dataset.